created test level component with reduced logic
